MHV-41341: Vaccine print view fixes
