+++ src/applications/mhv/medical-records/components/shared/ItemList.jsx
@@ -5,10 +5,10 @@ const ItemList = props => {
   const { list, emptyMessage } = props;
   if (list?.length > 1) {
     return (
-      <ul className="vads-u-margin-top--1">
+      <ul className="vads-u-margin-top--0 item-list">
         {list.map((item, idx) => {
           return (
-            <li key={idx} className="vads-u-margin-bottom-0">
+            <li key={idx} className="vads-u-margin-bottom--0">
               {item}
             </li>
           );
